+++ src/components/books/Books.jsx
@@ -18,7 +18,7 @@ export default function Books({sectionName}) {
       {/* BUTTON */}
       <TagsNav />
       <Welcome section={sectionName}/>
-      <Outlet></Outlet>
+      <Outlet key={pathname}></Outlet>
     </div>
   );
 }
