+++ src/components/challenges/Challenges.jsx
@@ -18,7 +18,7 @@ export default function Challenges({sectionName}) {
       {/* categories */}
       <TagsNav />
       <Welcome section={sectionName} />
-      <Outlet></Outlet>
+      <Outlet key={pathname}></Outlet>
     </div>
   );
 }
